requirements.txt
anyio==3.6.1
asyncpg==0.25.0
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.2
black==21.12b0
cffi==1.15.1
cfgv==3.3.1
click==8.1.3
colorama==0.4.4
coverage==6.2
databases==0.6.0
distlib==0.3.4
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.78.0
filelock==3.4.2
flake8==4.0.1
freezegun==1.1.0
greenlet==1.1.2
h11==0.13.0
identify==2.4.3
idna==3.3
iniconfig==1.1.1
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.16.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pyflakes==2.4.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-freezegun==0.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
uvicorn==0.18.1
virtualenv==20.13.0
